requirements.txt
fastapi
uvicorn[standard]
websockets
supabase
python-dotenv
